requirements.txt
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.4

Pillow==10.3.0
opencv-python-headless==4.10.0.84

torch==2.2.2
torchvision==0.17.2

transformers==4.41.2
ultralytics==8.3.0

autogluon.tabular==1.1.1